backend/generated_projects/web-provide-online-874670/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
sqlalchemy==2.0.31
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.3
